NMS-13104: Move getBooleanWithDefaultAsTrue to Sysprops
