+++ core/ipc/common/kafka/src/main/java/org/opennms/core/ipc/common/kafka/KafkaTopicProvider.java
@@ -33,9 +33,8 @@ import static org.opennms.core.ipc.common.kafka.KafkaRpcConstants.RPC_REQUEST_TO
 import static org.opennms.core.ipc.common.kafka.KafkaRpcConstants.RPC_RESPONSE_TOPIC_NAME;
 import static org.opennms.core.ipc.common.kafka.KafkaRpcConstants.SINGLE_TOPIC_FOR_ALL_MODULES;
 
+import org.opennms.core.sysprops.SystemProperties;
 import org.opennms.core.utils.SystemInfoUtils;
-
-import com.google.common.base.Strings;
 
 public class KafkaTopicProvider {
 
@@ -50,7 +49,7 @@ public class KafkaTopicProvider {
     }
 
     public KafkaTopicProvider() {
-        this.singleTopic = getBooleanWithDefaultAsTrue(String.format("%s%s", KAFKA_RPC_CONFIG_SYS_PROP_PREFIX, SINGLE_TOPIC_FOR_ALL_MODULES));
+        this.singleTopic = SystemProperties.getBooleanWithDefaultAsTrue(String.format("%s%s", KAFKA_RPC_CONFIG_SYS_PROP_PREFIX, SINGLE_TOPIC_FOR_ALL_MODULES));
     }
 
     public String getRequestTopicAtLocation(String location, String module) {
@@ -65,20 +64,6 @@ public class KafkaTopicProvider {
             return String.format(TOPIC_NAME_WITHOUT_LOCATION, SystemInfoUtils.getInstanceId(), RPC_RESPONSE_TOPIC_NAME);
         }
         return String.format(TOPIC_NAME_WITH_MODULE, SystemInfoUtils.getInstanceId(), RPC_RESPONSE_TOPIC_NAME, module);
-    }
-
-    /* Overwrite getBoolean to make default value to true */
-    public static Boolean getBooleanWithDefaultAsTrue(String name) {
-        boolean defaultValue = true;
-        try {
-            String singleTopic = System.getProperty(name);
-            if(!Strings.isNullOrEmpty(singleTopic)) {
-                defaultValue = !singleTopic.equalsIgnoreCase("false");
-            }
-        } catch (NullPointerException | IllegalArgumentException e) {
-            // Ignore
-        }
-        return defaultValue;
     }
 
 }
